test(multicloud): move & adjust `TestMulticloudIBMAzure` to `mutlcloud_test.go`
